Simplify the astar_search header
I moved the 'maze' type the .cpp file.
It did mean that 'make_maze' *has* to return a 'shared_ptr' if the pointer it returned was later going to be wrapped in a 'shared_ptr' which was going to happen anyway.
